more current sense debugging - attempt voltage-based position control ( i think the velocity reading is too noisy?? )

+++ include/config.h
@@ -32,7 +32,7 @@
 
 struct float_frame {
   String name;
-  std::vector<float> values;
+  std::vector<double> values;
 };
 
 struct string_frame {
